+++ test/decorators.js
@@ -1,9 +1,12 @@
-import chai from 'chai';
+import chai, { expect } from 'chai';
 import linkifyIt from 'linkify-it';
-import { ContentBlock } from 'draft-js';
+import { ContentBlock, convertFromRaw } from 'draft-js';
 import tlds from 'tlds';
 import redraft from '../src';
 import { joinRecursively } from './helpers';
+import TestDecorator from './TestDecorator';
+
+chai.should();
 
 const linkify = linkifyIt();
 linkify.tlds(tlds);
@@ -19,6 +22,10 @@ const linkStrategy = (contentBlock, callback) => {
 
 
 const link = ({ decoratedText, children }) => `<a href="${decoratedText}" >${joinRecursively(children)}</a>`;
+const linkWithContentState = ({ decoratedText, children, contentState }) => {
+  expect(contentState).to.be.an('object');
+  return `<a href="${decoratedText}" >${joinRecursively(children)}</a>`;
+};
 
 
 const decorators = [
@@ -28,8 +35,12 @@ const decorators = [
   },
 ];
 
-
-chai.should();
+const decoratorsContentState = [
+  {
+    strategy: linkStrategy,
+    component: linkWithContentState,
+  },
+];
 
 const rawWithLink = {
   entityMap: {},
@@ -57,6 +68,19 @@ const rawWithLink2 = {
   }],
 };
 
+const rawWithNoText = {
+  entityMap: {},
+  blocks: [{
+    key: '8ofc8',
+    text: '',
+    type: 'unstyled',
+    depth: 0,
+    inlineStyleRanges: [],
+    entityRanges: [],
+    data: {},
+  }],
+};
+
 
 const inline = {
   BOLD: (children) => `<strong>${children.join('')}</strong>`,
@@ -78,6 +102,14 @@ const renderers = {
   decorators,
 };
 
+const renderersWithContentState = {
+  inline,
+  blocks,
+  entities,
+  decorators: decoratorsContentState,
+};
+
+
 describe('redraft with decorators', () => {
   it('should apply decorator ranges and call decorator component', () => {
     const rendered = redraft(rawWithLink, renderers);
@@ -95,5 +127,26 @@ describe('redraft with decorators', () => {
     });
     const joined = joinRecursively(rendered);
     joined.should.equal('<a href="http://lokiuz.github.io/redraft/" >http://lokiuz.<strong>github</strong>.io/redraft/</a>'); // eslint-disable-line max-len
+  });
+  it('should handle convertFromRawToDraftState in options', () => {
+    const rendered = redraft(rawWithLink, renderersWithContentState, {
+      convertFromRaw,
+    });
+    const joined = joinRecursively(rendered);
+    joined.should.equal('<a href="http://lokiuz.github.io/redraft/" >http://lokiuz.<strong>github</strong>.io/redraft/</a>'); // eslint-disable-line max-len
+  });
+  it('should handle Decorator in options', () => {
+    const rendered = redraft(rawWithLink, renderers, {
+      Decorator: TestDecorator,
+    });
+    const joined = joinRecursively(rendered);
+    joined.should.equal(`<span style="first first-${rawWithLink.blocks[0].key}" >h</span>ttp://lokiuz.<strong>github</strong>.io/redraft/`); // eslint-disable-line max-len
+  });
+  it('should handle Decorator in options with empty block', () => {
+    const rendered = redraft(rawWithNoText, renderers, {
+      Decorator: TestDecorator,
+    });
+    const joined = joinRecursively(rendered);
+    joined.should.equal(''); // eslint-disable-line max-len
   });
 });